requirements.txt
langchain
langchain-community 
huggingface_hub
python-dotenv
langchain-huggingface
langchain-groq
grandalf
pydantic